refactor : LineSteps - 요청과 검증 Steps 으로 분리
