+++ src/test/java/nextstep/subway/line/LineRequestSteps.java
@@ -4,15 +4,9 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.line.dto.LineRequest;
-import nextstep.subway.line.dto.LineResponse;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class LineSteps {
+public class LineRequestSteps {
 
     public static ExtractableResponse<Response> 지하철_노선_생성_요청(String name, String color) {
         return RestAssured
@@ -25,19 +19,6 @@ public class LineSteps {
                 .extract();
     }
 
-    public static void 지하철_노선_생성_됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        생성된_지하철_노선_URI_경로_존재_함(response);
-    }
-
-    public static void 생성된_지하철_노선_URI_경로_존재_함(ExtractableResponse<Response> response) {
-        assertThat(생성된_지하철_노선_URI_경로_확인(response)).isNotBlank();
-    }
-
-    public static String 생성된_지하철_노선_URI_경로_확인(ExtractableResponse<Response> response) {
-        return response.header("Location");
-    }
-
     public static ExtractableResponse<Response> 지하철_노선_목록_조회_요청() {
         return RestAssured
                 .given().log().all()
@@ -45,15 +26,6 @@ public class LineSteps {
                 .get("/lines")
                 .then().log().all()
                 .extract();
-    }
-
-    public static void 지하철_노선_목록_조회_됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 지하철_노선_목록_조회_결과에_2개_노선_포함_확인(ExtractableResponse<Response> readLinesResponse) {
-        List<LineResponse> lines = readLinesResponse.jsonPath().getList(".", LineResponse.class);
-        assertThat(lines).hasSize(2);
     }
 
     public static ExtractableResponse<Response> 지하철_노선_조회_요청(String uri) {
@@ -64,10 +36,6 @@ public class LineSteps {
                 .get(uri)
                 .then().log().all()
                 .extract();
-    }
-
-    public static void 지하철_노선_조회_됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
     }
 
     public static ExtractableResponse<Response> 지하철_노선_수정_요청(Long id, String name, String color) {
@@ -82,14 +50,6 @@ public class LineSteps {
                 .extract();
     }
 
-    public static Long 수정할_지하철_노선_ID(ExtractableResponse<Response> createResponse) {
-        return createResponse.as(LineResponse.class).getId();
-    }
-
-    public static void 지하철_노선_수정_됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
     public static ExtractableResponse<Response> 지하철_노선_제거_요청(String uri) {
         return RestAssured
                 .given().log().all()
@@ -98,9 +58,4 @@ public class LineSteps {
                 .then().log().all()
                 .extract();
     }
-
-    public static void 지하철_노선_제거_됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
-    }
-
 }
